% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_age}
\alias{calculate_age}
\title{Add Age Variable to Dataset}
\usage{
calculate_age(
  data,
  dob_col = "dateOfBirth",
  test_date_col = "Testdate",
  output_col = "age"
)
}
\arguments{
\item{data}{Data frame containing date columns.}

\item{dob_col}{Character. Name of date of birth column. Default is "dateOfBirth".}

\item{test_date_col}{Character. Name of test date column. Default is "Testdate".}

\item{output_col}{Character. Name for the new age column. Default is "age".}
}
\value{
Data frame with added age column.
}
\description{
Calculates age in years based on date of birth and test date.
Handles missing dates gracefully.
}
\details{
Calculate Age from Date of Birth and Test Date
}
\keyword{internal}
